feat: spawn netcat tcp child process in debugAdapter
when connecting our debugAdapter to an socket (localhost:9999) also
spawn a rudimentary "server" by executing 'nc -l 9999" and take
messages recieved on the server back to the client (our debugAdapter)
and log to console.

Future work is to replace our 'nc -l 9999' call with a 'snail --debug 9999'
(or something similar) call to launch snail in a debug mode that
will allow us to communicate breakpoint, step in, continue, etc.
requests from a debugging session running in our extension.