fix: snail language server calls snail path from extension settings
Snail Language Server now gets a path to snail via extension settings.

Known problems: the default value of this path is 'snail'. This is
making an assumption that the user HAS a valid snail executable
downloaded AND that the path to said snail executable is in their
PATH to execute it with no path specification.

Further work will involve
(1) getting a better default snail path value
based on the machine the extension is launched on
(2) checking that the snail path gathered or given is valid
(exists, has a -s flag) and giving a nice error message if
the answer is no