feat: move snail validation to extension.ts
We now validate the snail path when we activate our extension.
If the path given is bad, the extension stops activation and
shows an error message (but does not provide a link to settings).

However, when the user changes the settings, it does not
re-validate that path or re-activate the extension until we
developer reload the window.

If we want the extension to automaticaly check for changes in our
extension settings, that needs to be defined in our language
server, and in order to show a nice message from our language
server, we need to figure out why server.ts doesn't find the
vscode module when our extension runs.