docs: update README for (almost) package release
Update README to contain important documentation about the
extension before release. Includes adding a list of features,
requirements for using the extension and extension settings.

Future work for this documentation includes refining how we
document our snippets autocompletions. Also need to include
a link to GitHub Issues page for when users experience issues
with the extension. Would also be nice to switch out our image
of static error checking with a dark theme image, or possibly
even an animation.